Add CORS headers with bottle_cors_plugin
Added dependency on Bottle plugin `bottle_cors_plugin`.

The plugin is "installed" into the Bottle app as part of the app
creation/config. We also needed to make a few adjustments to the JSON
error handling.

The allowed origin (`CORS_ORIGIN`) is currently hardcoded to
`localhost:5173`, which is our frontend dev server from Vite. This will
need to be loaded from an environment variable or similar in the
production deployment.