Change frontend Vite config to listen on port 8000
Changes Vite config to default the dev server to listening on port 8000,
instead of the Vite default of 5173. This applies when running the Vite
dev server directly (`npm run dev`) as well as in the Docker Compose
stack.

This also changes the frontend's host and port to be configurable by
environment variables `LISTEN_HOST` and `LISTEN_PORT`, and configures
these in the `docker-compose.yml`.

To match these changes, the backend's CORS headers now default to
`http://localhost:8000`, since this should be the expected value for all
dev setups.